+++ packages/syft/tests/conftest.py
@@ -72,4 +72,5 @@ pytest_plugins = [
     "tests.syft.users.fixtures",
     "tests.syft.metadata.fixtures",
     "tests.syft.dataset.fixtures",
+    "tests.syft.messages.fixtures",
 ]
