+++ packages/syft/tests/syft/messages/message_service_test.py
@@ -1,5 +1,4 @@
 # third party
-import pytest
 from pytest import MonkeyPatch
 from result import Err
 from result import Ok
@@ -19,58 +18,12 @@ from syft.core.node.new.messages import MessageStatus
 from syft.core.node.new.response import SyftError
 from syft.core.node.new.response import SyftSuccess
 from syft.core.node.new.uid import UID
-from syft.core.node.new.user import User
-from syft.core.node.worker import Worker
 
 test_verify_key_string = (
     "e143d6cec3c7701e6ca9c6fb83d09e06fdad947742126831a684a111aba41a8c"
 )
 
 test_verify_key = SyftVerifyKey.from_string(test_verify_key_string)
-
-
-@pytest.fixture(autouse=True)
-def message_stash(document_store):
-    return MessageStash(store=document_store)
-
-
-@pytest.fixture(autouse=True)
-def message_service(document_store):
-    return MessageService(store=document_store)
-
-
-@pytest.fixture(autouse=True)
-def authed_context(admin_user: User, worker: Worker) -> AuthedServiceContext:
-    return AuthedServiceContext(credentials=test_verify_key, node=worker)
-
-
-@pytest.fixture(autouse=True)
-def linked_object():
-    return LinkedObject(
-        node_uid=UID(),
-        service_type=MessageService,
-        object_type=Message,
-        object_uid=UID(),
-    )
-
-
-@pytest.fixture(autouse=True)
-def mock_create_message(faker) -> CreateMessage:
-    test_signing_key1 = SyftSigningKey.generate()
-    test_verify_key1 = test_signing_key1.verify_key
-    test_signing_key2 = SyftSigningKey.generate()
-    test_verify_key2 = test_signing_key2.verify_key
-
-    mock_message = CreateMessage(
-        subject="mock_created_message",
-        id=UID(),
-        node_uid=UID(),
-        from_user_verify_key=test_verify_key1,
-        to_user_verify_key=test_verify_key2,
-        created_at=DateTime.now(),
-    )
-
-    return mock_message
 
 
 def add_mock_message(
@@ -91,7 +44,7 @@ def add_mock_message(
         status=message_status_undelivered,
     )
 
-    result = message_stash.partition.set(mock_message)
+    result = message_stash.set(mock_message)
     assert result.is_ok()
 
     return mock_message
